+++ tsz.go
@@ -21,8 +21,8 @@ type Series struct {
 	sync.Mutex
 
 	// TODO(dgryski): timestamps in the paper are uint64
-	T0  uint32
-	t   uint32
+	T0  uint64
+	t   uint64
 	val float64
 
 	bw       bstream
@@ -30,11 +30,11 @@ type Series struct {
 	trailing uint8
 	finished bool
 
-	tDelta uint32
+	tDelta uint64
 }
 
 // New series
-func New(t0 uint32) *Series {
+func New(t0 uint64) *Series {
 	s := Series{
 		T0:      t0,
 		leading: ^uint8(0),
@@ -72,7 +72,7 @@ func (s *Series) Finish() {
 }
 
 // Push a timestamp and value to the series
-func (s *Series) Push(t uint32, v float64) {
+func (s *Series) Push(t uint64, v float64) {
 	s.Lock()
 	defer s.Unlock()
 
@@ -159,9 +159,9 @@ func (s *Series) Iter() *Iter {
 
 // Iter lets you iterate over a series.  It is not concurrency-safe.
 type Iter struct {
-	T0 uint32
+	T0 uint64
 
-	t   uint32
+	t   uint64
 	val float64
 
 	br       bstream
@@ -170,7 +170,7 @@ type Iter struct {
 
 	finished bool
 
-	tDelta uint32
+	tDelta uint64
 	err    error
 }
 
@@ -184,7 +184,7 @@ func bstreamIterator(br *bstream) (*Iter, error) {
 	}
 
 	return &Iter{
-		T0: uint32(t0),
+		T0: uint64(t0),
 		br: *br,
 	}, nil
 }
@@ -208,7 +208,7 @@ func (it *Iter) Next() bool {
 			it.err = err
 			return false
 		}
-		it.tDelta = uint32(tDelta)
+		it.tDelta = uint64(tDelta)
 		it.t = it.T0 + it.tDelta
 		v, err := it.br.readBits(64)
 		if err != nil {
@@ -276,7 +276,7 @@ func (it *Iter) Next() bool {
 		dod = int32(bits)
 	}
 
-	tDelta := it.tDelta + uint32(dod)
+	tDelta := it.tDelta + uint64(dod)
 
 	it.tDelta = tDelta
 	it.t = it.t + it.tDelta
@@ -335,7 +335,7 @@ func (it *Iter) Next() bool {
 }
 
 // Values at the current iterator position
-func (it *Iter) Values() (uint32, float64) {
+func (it *Iter) Values() (uint64, float64) {
 	return it.t, it.val
 }
